requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==26.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
cryptography==46.0.6
    # via pyjwt
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via mcp
httpx-sse==0.4.3
    # via mcp
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
mcp==1.26.0
    # via agentmux (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
packaging==26.0
    # via pytest
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via questionary
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pygments==2.20.0
    # via
    #   pytest
    #   rich
pyjwt[crypto]==2.12.1
    # via mcp
pytest==9.0.2
    # via agentmux (pyproject.toml)
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.22
    # via mcp
pyyaml==6.0.3
    # via agentmux (pyproject.toml)
questionary==2.1.1
    # via agentmux (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==14.3.3
    # via agentmux (pyproject.toml)
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shtab==1.8.0
    # via agentmux (pyproject.toml)
sse-starlette==3.3.4
    # via mcp
starlette==1.0.0
    # via
    #   mcp
    #   sse-starlette
typing-extensions==4.15.0
    # via
    #   anyio
    #   mcp
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
uvicorn==0.42.0
    # via mcp
watchdog==6.0.0
    # via agentmux (pyproject.toml)
wcwidth==0.6.0
    # via prompt-toolkit
